Avoid race condition when renewing certificates
When renewing several certificates, service could fail to reload
properly due to a race condition, if the reload happens while lecm
is re-writing the pem file.

By default, subprocess.Popen() is not blocking, to ensure we do not face
this race condition again, we make it blocking, by waiting for the
process to terminante.