Do not clean filesystem before requesting certs
The .csr, .crt and .pem files are currently removed before requesting
the certificate from Let's Encrypt. This can lead to cases where - for
any reason - the certificate is not returned, and then .crt and .pem
files are lost.

All three of them are properly (re)generated during the
_create_certificate() function, overriding if necessary currently
present files.