Run the reload command only when necessary
Before this patch the reload command was run for each and every new
certificate generation or renewal even if all certificates were used
for the same service.

From now it will reload the services only when necessary, so if all
services uses the same service_name then it will be reload only once.