Get the intermediate certificate whenever needed
The intermediate certificate was currently retrieved locally only for
the 1st certificate generation.

This causes issues when the intermediate cerficiate changes (recent
change from X3 to R3) and the certificates just need to be renewed.

This patch moves the retrieval of the intermediate certificate along
with the certificate creation to be sure the former exists before being
attached to the latter.